hw/audio/opl2lpt: add support for OPL2LPT
OPL2LPT is an OPL2 chip on the parallel port. It can be programmed
mostly like an OPL2 chip, except it is write-only. Timers are
therefore not usable and we emulate a simplified version of them: they
expire immediately (in the future, this could be updated to properly
emulate them). Timers are usually only used to detect the card.

The driver needs a parallel port to be usable:

    -chardev parport,id=parport0,path=/dev/parport0
    -device opl2lpt,chardev=parport0

From a timing perspective, usleep() is used to ensure to give enough
time to OPL2 to react properly. Ideally, the timing should be handled
by the game, but testing with Indy3, this didn't work. I have also
tried to be smarter by sleeping less depending on the current clock
but a few glitches were still present. With the current approach,
Indy3 is able to use the sound card without a glitch. The DOS driver
doesn't add delays (but it doesn't work on QEMU either due to similar
issues, so it may be a clue).

On shutdown, the OPL2 chip should be reset to not continue sustaining
a note. I was unable to find the appropriate function to
register (unrealize callback is not called on shutdown and
qemu_register() isn't always called on shutdown either). The user has
to reset the VM, then stop it if they don't want to reset the OPL2LPT
manually.

Signed-off-by: Vincent Bernat <vincent@bernat.im>